# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Third Arm API
  version: "0.1.0"
  description: |
    REST northbound contract for the Third Arm intent-driven handover service.
    Stage 1 — operator-triggered, mock hardware.
  contact:
    name: Third Arm Team

# TODO: replace localhost with real base URL before staging/production deploy
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /session/start:
    post:
      tags: [session]
      summary: Begin a new handover session
      operationId: startSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionStartRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "409":
          description: Session already active
  /session/stop:
    post:
      tags: [session]
      summary: End or abort current session
      operationId: stopSession
      responses:
        "200":
          description: Session stopped
        "404":
          description: No active session
...
components:
  schemas:
...
    SessionStartRequest:
      type: object
      properties:
        operator_id: { type: string, example: "operator-01" }
        object_id:   { type: string, example: "obj_water_bottle_500ml" }
        slot_id:     { type: string, example: "slot_A" }
        notes:       { type: string }
    SessionResponse:
      type: object
      properties:
        session_id: { type: string }
        started_at: { type: string, format: date-time }
...